add basic config
