+++ protocol/messenger.go
@@ -3,7 +3,6 @@ package protocol
 import (
 	"context"
 	"crypto/ecdsa"
-	"database/sql"
 	"io/ioutil"
 	"math/rand"
 	"os"
@@ -26,6 +25,7 @@ import (
 	"github.com/status-im/status-go/protocol/images"
 	"github.com/status-im/status-go/protocol/protobuf"
 	"github.com/status-im/status-go/protocol/push_notification_client"
+	"github.com/status-im/status-go/protocol/push_notification_server"
 	"github.com/status-im/status-go/protocol/sqlite"
 	"github.com/status-im/status-go/protocol/transport"
 	wakutransp "github.com/status-im/status-go/protocol/transport/waku"
@@ -59,6 +59,7 @@ type Messenger struct {
 	processor                  *messageProcessor
 	handler                    *MessageHandler
 	pushNotificationClient     *push_notification_client.Client
+	pushNotificationServer     *push_notification_server.Server
 	logger                     *zap.Logger
 	verifyTransactionClient    EthClient
 	featureFlags               featureFlags
@@ -100,99 +101,6 @@ type featureFlags struct {
 type dbConfig struct {
 	dbPath string
 	dbKey  string
-}
-
-type config struct {
-	// This needs to be exposed until we move here mailserver logic
-	// as otherwise the client is not notified of a new filter and
-	// won't be pulling messages from mailservers until it reloads the chats/filters
-	onNegotiatedFilters func([]*transport.Filter)
-	// DEPRECATED: no need to expose it
-	onSendContactCodeHandler func(*encryption.ProtocolMessageSpec)
-
-	// systemMessagesTranslations holds translations for system-messages
-	systemMessagesTranslations map[protobuf.MembershipUpdateEvent_EventType]string
-	// Config for the envelopes monitor
-	envelopesMonitorConfig *transport.EnvelopesMonitorConfig
-
-	messagesPersistenceEnabled bool
-	featureFlags               featureFlags
-
-	// A path to a database or a database instance is required.
-	// The database instance has a higher priority.
-	dbConfig dbConfig
-	db       *sql.DB
-
-	verifyTransactionClient EthClient
-
-	logger *zap.Logger
-}
-
-type Option func(*config) error
-
-// WithSystemMessagesTranslations is required for Group Chats which are currently disabled.
-// nolint: unused
-func WithSystemMessagesTranslations(t map[protobuf.MembershipUpdateEvent_EventType]string) Option {
-	return func(c *config) error {
-		c.systemMessagesTranslations = t
-		return nil
-	}
-}
-
-func WithOnNegotiatedFilters(h func([]*transport.Filter)) Option {
-	return func(c *config) error {
-		c.onNegotiatedFilters = h
-		return nil
-	}
-}
-
-func WithCustomLogger(logger *zap.Logger) Option {
-	return func(c *config) error {
-		c.logger = logger
-		return nil
-	}
-}
-
-func WithMessagesPersistenceEnabled() Option {
-	return func(c *config) error {
-		c.messagesPersistenceEnabled = true
-		return nil
-	}
-}
-
-func WithDatabaseConfig(dbPath, dbKey string) Option {
-	return func(c *config) error {
-		c.dbConfig = dbConfig{dbPath: dbPath, dbKey: dbKey}
-		return nil
-	}
-}
-
-func WithVerifyTransactionClient(client EthClient) Option {
-	return func(c *config) error {
-		c.verifyTransactionClient = client
-		return nil
-	}
-}
-
-func WithDatabase(db *sql.DB) Option {
-	return func(c *config) error {
-		c.db = db
-		return nil
-	}
-}
-
-func WithDatasync() func(c *config) error {
-	return func(c *config) error {
-		c.featureFlags.datasync = true
-		return nil
-	}
-}
-
-func WithEnvelopesMonitorConfig(emc *transport.EnvelopesMonitorConfig) Option {
-	return func(c *config) error {
-		c.envelopesMonitorConfig = emc
-		return nil
-	}
 }
 
 func NewMessenger(
@@ -333,6 +241,12 @@ func NewMessenger(
 	pushNotificationClientPersistence := push_notification_client.NewPersistence(database)
 	pushNotificationClient := push_notification_client.New(pushNotificationClientPersistence)
 
+	var pushNotificationServer *push_notification_server.Server
+	if c.pushNotificationServerConfig != nil {
+		pushNotificationServerPersistence := push_notification_server.NewSQLitePersistence(database)
+		pushNotificationServer = push_notification_server.New(c.pushNotificationServerConfig, pushNotificationServerPersistence)
+	}
+
 	processor, err := newMessageProcessor(
 		identity,
 		database,
@@ -357,6 +271,7 @@ func NewMessenger(
 		processor:                  processor,
 		handler:                    handler,
 		pushNotificationClient:     pushNotificationClient,
+		pushNotificationServer:     pushNotificationServer,
 		featureFlags:               c.featureFlags,
 		systemMessagesTranslations: c.systemMessagesTranslations,
 		allChats:                   make(map[string]*Chat),
